+++ complexOrm/src/main/kotlin/com/github/silasgermany/complexorm/ComplexOrmQueryBuilder.kt
@@ -15,7 +15,7 @@ class ComplexOrmQueryBuilder internal constructor(private val complexOrmReader: 
     private val KClass<out ComplexOrmTable>.tableName get() = complexOrmTableInfo.basicTableInfo.getValue(java.canonicalName!!).first
 
     private val restrictions = mutableMapOf<String, String>()
-    private val existingEntries = mutableMapOf<String, MutableMap<Int, ComplexOrmTable>>()
+    private val existingEntries = mutableMapOf<String, MutableMap<Long, ComplexOrmTable>>()
 
     private inline fun <reified T : ComplexOrmTable> specialWhere(
         selection: String,
@@ -124,8 +124,8 @@ class ComplexOrmQueryBuilder internal constructor(private val complexOrmReader: 
     override fun toString() = restrictions.values
             .joinToString(" AND ").replace("$$.", "")
 
-    inline fun <reified T : ComplexOrmTable> get(id: Int?): T? = get(T::class, id)
-    fun <T : ComplexOrmTable> ComplexOrmQueryBuilder.get(table: KClass<T>, id: Int?): T? {
+    inline fun <reified T : ComplexOrmTable> get(id: Long?): T? = get(T::class, id)
+    fun <T : ComplexOrmTable> ComplexOrmQueryBuilder.get(table: KClass<T>, id: Long?): T? {
         id ?: return null
         val tableClassName = table.java.canonicalName!!
         this@ComplexOrmQueryBuilder.restrictions[tableClassName] = if (tableClassName !in restrictions) "$$.id = $id"
