+++ complexOrm/src/main/kotlin/com/github/silasgermany/complexorm/models/ReadTableInfo.kt
@@ -7,7 +7,7 @@ import java.io.File
 
 class ReadTableInfo constructor(
         val restrictions: Map<String, String>,
-        private val alreadyLoaded: MutableMap<String, MutableMap<Int, ComplexOrmTable>>,
+        private val alreadyLoaded: MutableMap<String, MutableMap<Long, ComplexOrmTable>>,
         private val complexOrmTableInfo: ComplexOrmTableInfoInterface
 ) {
     var readIndex = 0
@@ -32,9 +32,9 @@ class ReadTableInfo constructor(
     fun alreadyGiven(tableClassName: String) = tableClassName in givenTables && !isMissingRequest(tableClassName)
     fun alreadyLoading(tableClassName: String) = tableClassName in loadingTables
     fun has(tableClassName: String) = alreadyLoaded.containsKey(tableClassName)
-    fun getTable(tableClassName: String, id: Int?) = alreadyLoaded[tableClassName]?.get(id)
+    fun getTable(tableClassName: String, id: Long?) = alreadyLoaded[tableClassName]?.get(id)
     fun setTable(tableClassName: String, table: ComplexOrmTable, specialColumnValue: String? = null) {
-        (table.map[specialColumnValue ?: "id"] as Int?)?.let { alreadyLoaded.init(tableClassName)[it] = table }
+        (table.map[specialColumnValue ?: "id"] as Long?)?.let { alreadyLoaded.init(tableClassName)[it] = table }
     }
     fun isMissingRequest(tableClassName: String) = missingEntries?.any { it.javaClass.canonicalName == tableClassName } == true
     fun print() {
